refactor: move last solved to stats page
