+++ helper.js
@@ -1,3 +1,6 @@
+const crypto = require('crypto')
+const secrets = require('./secrets-loader.js')
+
 function capitalizeFirstLetter(inputString) {
   // Check if the input is a non-empty string
   if (typeof inputString !== 'string' || inputString.length === 0) {
@@ -8,4 +11,16 @@ function capitalizeFirstLetter(inputString) {
   return inputString.charAt(0).toUpperCase() + inputString.slice(1)
 }
 
-module.exports = { capitalizeFirstLetter }
+function generateSHA256(input) {
+  const hash = crypto.createHash('sha256')
+  hash.update(input)
+  return hash.digest('hex')
+}
+
+function generateToken(userId) {
+  return `${userId}-${generateSHA256(
+    `${userId}___${secrets('config_token_secret')}`
+  ).substring(0, 12)}`
+}
+
+module.exports = { capitalizeFirstLetter, generateSHA256, generateToken }
